Moved TodoGraphQLWithAPIKeyTests over to GraphQLWithUserPoolIntegrationTests since they use the same schema
